Updated copyright helper script, in order to work with git.

+++ tools/copyright_update_helper.sh
@@ -1,42 +1,74 @@
-############# IMPORTANT NOTE ###############
-# This script is not up to date anymore since when we switched from svn to git to manage Sollya
-# It must be updated.
-
 #!/bin/sh
 
-start_rev=2023
+# Think also of updating sollya.tex, sollya.php (two places)
+# and the meta copyright emitLegalNoticeAndDisclaimer in
+# implement.c, as well as the corresponding reference files
+# in the tests.
+
+# This script explicitly exclude the following files, that must
+# hence be manually checked:
+#    general.c
+#    help.h
+#    implement.c
+#    xml.c
+#    parser.y
+
+# At the time of designing this script, only files of the form
+# *.h *.c *.cpp *.y *.l are copyrighted. Also, no file in the
+# subdirectories, except the above mentioned sollya.[tex|php] are
+# copyrighted. But it might be good to check that it is still the
+# case every time this script is run.
+
+start_rev=sollya-4.1~
 end_rev=HEAD
+
+# A space separated list of hashes can be provided below. They will
+# be excluded from the search. Typically useful when a particular
+# commit changed a comma in all files of the repository
 excluded_revs=""
 
+# Methodology: the file /tmp/sollya-copyright/copyright.log contains an entry
+# for each copyrighted file in Sollya. This entry first sums up the current
+# copyright and contributors notice. Then, it lists all commits where the file
+# has been modified since start_rev, together with the name of the contributor
+# who performs the commit and the year of the commit.
+
+# Note: we list years of the copyright as a range year1-year2, where year1 is
+# the year of creation of the file, and year2 is the year of last modification,
+# even if the file has not been modified during some year in-between.
+
 cd /tmp
-svn co svn+ssh://schevill@scm.gforge.inria.fr/svn/sollya/trunk sollya-copyright
+git clone https://scm.gforge.inria.fr/anonscm/git/sollya/sollya.git sollya-copyright
 cd sollya-copyright
 
 if [ -z "$excluded_revs" ]
-  then excluded_revs=`expr $start_rev - 1`
+  then excluded_revs=ffffffffffffffffffffffffffffffffffffffff
 fi
 
-for file in `grep --recursive -i -l "Copyright.*20" * | grep -v "\.svn"`
+for file in *.c *.h *.cpp *.l *.y
 do
-  svn log $file -r$start_rev:$end_rev | grep "^r[0-9]" | sed -n 's/^\(r[0-9]* | [^ ]* | [0-9]*\).*/\1/;p' | grep -v `printf "$excluded_revs" |sed -n 's/ /\\\\|r/g;p' |sed -n 's/^/r/;p'` | sed -n 's/^/   /;p' > __sollya_tmp
-  if grep "^[[:space:]]*r[0-9]" __sollya_tmp > /dev/null
+  # WARNING: due to their particular content, these files
+  # must be manually handled.
+  if [ $file != "general.c" -a $file != "help.h" -a $file != "implement.c" -a $file != "xml.c" -a $file != "parser.y" ]
   then
-    notice=`grep -i Copyright $file`
-    notice=$notice" "`grep -i Contributor $file`
-    printf "%s" "$file:"
-    printf "\n"
-    printf "%s" "$notice"
-    printf "\n"
-    cat __sollya_tmp
-    printf "\n\n"
+    git log --pretty=format:'%H | %an | %ai' $start_rev..$end_rev $file | grep -v "^\("`printf "$excluded_revs" |sed -n 's/ /\\\\|/g;p'`"\)" | sed -n 's/^/   /;p' | sed -n 's/clauter/Christoph Lauter/;p' | sed -n 's/schevill/Sylvain Chevillard/;p' | sed -n 's/\(20[0-9][0-9]\)-.*$/\1/;p' > __sollya_tmp
+
+    if grep "^[[:space:]]*[0-9a-f]" __sollya_tmp > /dev/null
+    then
+      notice=`grep -i Copyright $file`
+      notice=$notice" "`grep -i Contributor $file`
+      printf "%s" "$file:"
+      printf "\n"
+      printf "%s" "$notice"
+      printf "\n"
+      cat __sollya_tmp
+      printf "\n\n"
+    fi
+    rm __sollya_tmp
   fi
-  rm __sollya_tmp
 done > copyright.log
-
-# Be aware that there could be false positives. Sometimes, a commit does not require to change the copyright notice (typically, a change of syntax over the whole source code). Such revisions can be explicitely excluded by adding them to excluded_revs variable.
-# Also, be aware that the work done in branches is not reflected by copyright.log.
-
-# Methodology: the file /tmp/sollya-copyright/copyright.log contains an entry for each copyrighted file in Sollya. This entry first sums up the current copyright and contributors notice. Then, it lists all commits where the file has been modified since start_rev, together with the name of the contributor who performs the commit and the year of the commit.
-
-# Note: we list years of the copyright as a range year1-year2, where year1 is the year of creation of the file, and year2 is the year of last modification, even if the file has not been modified during some year in-between.
-
+mv copyright.log /tmp
+cd /tmp
+rm -rf sollya-copyright
+cat copyright.log
+rm copyright.log
